rename approval row file
